hope I fixed a can can problem
Former-commit-id: be3a73d61eb30882ee90a7c6d54104924f627b52 [formerly 310e4feda15d89add0f642a38f4f1b8a1f68686d] [formerly ec835dda0199fa3cd3d7b55d6358e3e72e3affc9 [formerly 8c9cbc6f397e581c28dfc85283ba758ff2ce5b5f]] [formerly d2a91decb1d51a518fbfc81c1a6fdc1952e06af1 [formerly 8045270de4eb6e40ad6f39262050f5334ff33462] [formerly 8f4f0059d954b96b1d7972a1766d7dfc5f385cdd [formerly 8f4f0059d954b96b1d7972a1766d7dfc5f385cdd [formerly 8a799280fe9d49b773f304a77c30368d81f57c15]]]]
Former-commit-id: 14ab13639230b9aded6c5abce1f0ea6eb8436b8c [formerly 0fd3afa38675f72ecd22c1b222a5cb7e7df2e88b] [formerly 97cb7ae519ae2f34d2104eb18b75266dd349e137 [formerly 8686d88f416ad557ba7e6f50d76ca4f7738f6892]]
Former-commit-id: 5b7371e9d193dc4460fccaf14792ccab78dfcd02 [formerly 4908e4f6420319e5d28693de648773f00983dbbb]
Former-commit-id: 07fa3f683a004b3a2509e5fa9db72ed94147a544
Former-commit-id: 4f55b8c76ca41eaa9f56b02da0ad2bada6e429ec

+++ @server/controllers/comment_controller.rb
@@ -84,11 +84,13 @@ class CommentController < ApplicationController
 
   def destroy
     comment = Comment.find params[:id]
-    authorize! :destroy, comment
-
-    dirty_key("/comments/#{comment.point_id}")
-    comment.destroy
-
+    if current_user.id == comment.user_id
+      dirty_key("/comments/#{comment.point_id}")
+      comment.destroy
+    else
+      raise 'No permission to delete that comment (how do I implement this with can can?)'
+    end
+    
     render :json => []
   end
